refactor: move pipeline/evaluation.rs into astra-turn-core (620 LOC)
The only non-pipeline dependency was astra-turn-core::chat_turn_heuristics::
looks_like_live_query_with_context. Since astra-turn-core already depends on
astra-pipeline, the module belongs in turn-core — no cycle created.

- astra-turn-core/src/evaluation.rs: new module (was runtime/src/pipeline/evaluation.rs)
- runtime/src/pipeline/evaluation.rs: 6-line re-export stub
- astra-turn-core/src/lib.rs: registers the new module
- No behavior changes

Stacks on PRs #124 and #125.
